finish implementation & tests
