+++ app/src/main/java/ch/epfl/sdp/cook4me/Cook4MeApp.kt
@@ -10,7 +10,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import ch.epfl.sdp.cook4me.persistence.model.Post
-import ch.epfl.sdp.cook4me.ui.OverviewScreen
+import ch.epfl.sdp.cook4me.ui.overview.OverviewScreen
 import ch.epfl.sdp.cook4me.ui.eventform.CreateEventScreen
 import ch.epfl.sdp.cook4me.ui.eventform.DetailedEventScreen
 import ch.epfl.sdp.cook4me.ui.eventform.Event
