+++ app/src/main/java/ch/epfl/sdp/cook4me/application/AccountService.kt
@@ -3,7 +3,6 @@ package ch.epfl.sdp.cook4me.application
 import android.util.Patterns
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
-import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.tasks.await
 
 class AccountService(private val auth: FirebaseAuth = FirebaseAuth.getInstance()) {
@@ -19,20 +18,8 @@ class AccountService(private val auth: FirebaseAuth = FirebaseAuth.getInstance()
         }
         return currentUser.email
     }
-    suspend fun signOut(): Result<String> {
-        return try {
-            auth.signOut()
-            /*
-            Just signing out actually won't throw any exceptions,
-            even when there is no user logged in.
-            However if we need to do something else upon signing out,
-            e.g. call another function like revokeAccessFromServer(),
-            which could throw an exception, then we should wrap it in a try-catch block.
-            */
-            Result.success("Sign out successful")
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
+    fun signOut() {
+        auth.signOut()
     }
     suspend fun authenticate(email: String, password: String) {
         auth.signInWithEmailAndPassword(email, password).await()
